backend/requirements.txt
asgiref
celery
django-grappelli
django-celery-results
certifi
chardet
Django
django-cors-headers
django-filter
django-simple-history
djangorestframework
django-debug-toolbar
django-import-export
django-url-filter
future
idna
mysqlclient
pytz
requests
six
sqlparse
urllib3
pandas
scrapy
python-dateutil
